requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
kagglehub
PyYAML
pyspark
onetl
fastapi
uvicorn